+++ core/src/com/unciv/models/ruleset/tile/TileImprovement.kt
@@ -55,7 +55,7 @@ class TileImprovement : RulesetStatsObject() {
 
     @Readonly fun isGreatImprovement() = hasUnique(UniqueType.GreatImprovement)
     @Readonly fun isRoad() = RoadStatus.entries.any { it != RoadStatus.None && it.name == this.name }
-    @Readonly fun isAncientRuinsEquivalent() = hasUnique(UniqueType.IsAncientRuinsEquivalent)
+    @Readonly fun isAncientRuinsEquivalent(state: GameContext? = GameContext.IgnoreConditionals) = hasUnique(UniqueType.IsAncientRuinsEquivalent, state)
 
     @Readonly fun canBeBuiltOn(terrain: String): Boolean = terrain in terrainsCanBeBuiltOn
     @Readonly fun canBeBuiltOn(terrain: Terrain): Boolean = terrainsCanBeBuiltOn.any { terrain.matchesFilter(it) }